+++ mfhdf/hrepack/hrepack_opttable.c
@@ -35,7 +35,7 @@ options_table_init(options_table_t **tbl)
 
     op_tbl->size   = 3;
     op_tbl->nelems = 0;
-    op_tbl->objs   = (pack_info_t *)malloc(op_tbl->size * sizeof(pack_info_t));
+    op_tbl->objs   = (pack_info_t *)malloc((size_t)op_tbl->size * sizeof(pack_info_t));
 
     for (i = 0; i < op_tbl->size; i++) {
         strcpy(op_tbl->objs[i].objpath, "\0");
@@ -82,7 +82,7 @@ options_add_chunk(obj_list_t *obj_list, int n_objs, int32 *chunk_lengths, int ch
 
     if (op_tbl->nelems + n_objs >= op_tbl->size) {
         op_tbl->size += n_objs;
-        op_tbl->objs = (pack_info_t *)realloc(op_tbl->objs, op_tbl->size * sizeof(pack_info_t));
+        op_tbl->objs = (pack_info_t *)realloc(op_tbl->objs, (size_t)op_tbl->size * sizeof(pack_info_t));
         for (i = op_tbl->nelems; i < op_tbl->size; i++) {
             strcpy(op_tbl->objs[i].objpath, "\0");
             op_tbl->objs[i].comp.info  = -1;
@@ -162,7 +162,7 @@ options_add_comp(obj_list_t *obj_list, int n_objs, comp_info_t comp, options_tab
 
     if (op_tbl->nelems + n_objs >= op_tbl->size) {
         op_tbl->size += n_objs;
-        op_tbl->objs = (pack_info_t *)realloc(op_tbl->objs, op_tbl->size * sizeof(pack_info_t));
+        op_tbl->objs = (pack_info_t *)realloc(op_tbl->objs, (size_t)op_tbl->size * sizeof(pack_info_t));
         for (i = op_tbl->nelems; i < op_tbl->size; i++) {
             strcpy(op_tbl->objs[i].objpath, "\0");
             op_tbl->objs[i].comp.info  = -1;
